[feat] 게임상태에서 턴을 갖고 있고, 이동시 턴을 변경하도록 변경
